Create separate organization routing module
