+++ src/app/organizations/layouts/organization-layout.component.html
@@ -1,4 +1,3 @@
-<app-navbar></app-navbar>
 <div class="org-nav" *ngIf="organization">
   <div class="container d-flex">
     <div class="d-flex flex-column">
@@ -57,4 +56,3 @@
   </div>
 </div>
 <router-outlet></router-outlet>
-<app-footer></app-footer>
